rework free-regions to be handled more specially
This may not be the most elegant way to do it, but it was the quickest
way I saw to get to a viable prototype. Free regions are still
considered points in the CFG, but they are not connected to the normal
CFG. Instead, during region inference, when we have an outlives
relationship `A: B @ P`, if we find that the end of the function is
reachable from `P` (within `B`), then we copy all skolemized-end points
found in `B` into `A`. We also initialize the "inference variable" for
free regions to contain the entire CFG + the appropriate skolemized
end-point.

We are still missing some bits. Primarily, we need to either no longer
use an inference variable for free regions (and instead use
a *constant*) or else put a "max" value on what a free-region can
be (i.e., if there are multiple free regions, we should not infer that
`'a` includes `end('b)` unless `'a: 'b` is declared -- if that is
otherwise required, it would be an error that indicates that the user
did something like `&'b *r_a` where `r_a` is something of type `&'a
T`).